pyproject.toml
[project]
name = "pet-adoption"
version = "0.1.0"
description = "Pet Adoption Platform"
authors = [
    {name = "Ivanovich"}
]
readme = "README.md"
requires-python = "^3.13.2"
dependencies = [
    "django>=5.0.1",
    "pillow>=11.1.0",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
